fix: CommentId にバリデーションを追加して、value を String に変更

+++ src/main/kotlin/com/example/realworldkotlinspringbootjdbc/presentation/response/Comment.kt
@@ -23,7 +23,7 @@ data class Comment(
     companion object {
         fun from(comment: com.example.realworldkotlinspringbootjdbc.domain.Comment): Comment =
             Comment(
-                comment.id.value,
+                comment.id.value.toInt(),
                 comment.body.value,
                 comment.createdAt,
                 comment.updatedAt,
